Refactor componentes del navbar
