+++ DREAM Lab Frontend/src/GlobalComponents/NavBar/NavBar.jsx
@@ -4,7 +4,7 @@ import UserAvatar from "./components/UserAvatar/UserAvatar";
 import Logout from "./components/Logout/Logout";
 import "../../App.css";
 import SearchBar from "../../components/general/SearchBar";
-import BotonCrearExperiencia from "../BotonCrearExperiencia/BotonCrearExperiencia";
+import BotonCrearExperiencia from "./components/BotonCrearExperiencia/BotonCrearExperiencia";
 
 import "./NavBar.css";
