+++ DREAM Lab Frontend/src/GlobalComponents/NavBar/components/BotonCrearExperiencia/BotonCrearExperiencia.jsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import ExperienciaIcon from "../assets/images/crearExperiencia.svg";
+import ExperienciaIcon from "../../assets/images/crearExperiencia.svg";
 import "./BotonCrearExperiencia.css";
 
 function BotonCrearExperiencia() {
